+++ src/backends/postgresql/standard-use-type.cpp
@@ -73,8 +73,7 @@ void postgresql_standard_use_type_backend::pre_use(indicator const * ind)
                 std::size_t const bufSize
                     = std::numeric_limits<int8_t>::digits10 + 3;
                 buf_ = new char[bufSize];
-                snprintf(buf_, bufSize, "%d",
-                    static_cast<int>(exchange_type_cast<x_int8>(data_)));
+                snprintf(buf_, bufSize, "%d", exchange_type_cast<x_int8>(data_));
             }
             break;
         case x_uint8:
@@ -82,8 +81,7 @@ void postgresql_standard_use_type_backend::pre_use(indicator const * ind)
                 std::size_t const bufSize
                     = std::numeric_limits<uint8_t>::digits10 + 3;
                 buf_ = new char[bufSize];
-                snprintf(buf_, bufSize, "%u",
-                    static_cast<unsigned int>(exchange_type_cast<x_uint8>(data_)));
+                snprintf(buf_, bufSize, "%u", exchange_type_cast<x_uint8>(data_));
             }
             break;
         case x_int16:
@@ -91,8 +89,7 @@ void postgresql_standard_use_type_backend::pre_use(indicator const * ind)
                 std::size_t const bufSize
                     = std::numeric_limits<int16_t>::digits10 + 3;
                 buf_ = new char[bufSize];
-                snprintf(buf_, bufSize, "%d",
-                    static_cast<int>(exchange_type_cast<x_int16>(data_)));
+                snprintf(buf_, bufSize, "%d", exchange_type_cast<x_int16>(data_));
             }
             break;
         case x_uint16:
@@ -100,8 +97,7 @@ void postgresql_standard_use_type_backend::pre_use(indicator const * ind)
                 std::size_t const bufSize
                     = std::numeric_limits<uint16_t>::digits10 + 3;
                 buf_ = new char[bufSize];
-                snprintf(buf_, bufSize, "%u",
-                    static_cast<unsigned int>(exchange_type_cast<x_uint16>(data_)));
+                snprintf(buf_, bufSize, "%u", exchange_type_cast<x_uint16>(data_));
             }
             break;
         case x_int32:
@@ -109,8 +105,7 @@ void postgresql_standard_use_type_backend::pre_use(indicator const * ind)
                 std::size_t const bufSize
                     = std::numeric_limits<int32_t>::digits10 + 3;
                 buf_ = new char[bufSize];
-                snprintf(buf_, bufSize, "%d",
-                    static_cast<int>(exchange_type_cast<x_int32>(data_)));
+                snprintf(buf_, bufSize, "%d", exchange_type_cast<x_int32>(data_));
             }
             break;
         case x_uint32:
@@ -118,8 +113,7 @@ void postgresql_standard_use_type_backend::pre_use(indicator const * ind)
                 std::size_t const bufSize
                     = std::numeric_limits<uint32_t>::digits10 + 3;
                 buf_ = new char[bufSize];
-                snprintf(buf_, bufSize, "%u",
-                    static_cast<unsigned int>(exchange_type_cast<x_uint32>(data_)));
+                snprintf(buf_, bufSize, "%u", exchange_type_cast<x_uint32>(data_));
             }
             break;
         case x_int64:
