+++ src/backends/postgresql/vector-use-type.cpp
@@ -108,8 +108,7 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                     std::size_t const bufSize
                         = std::numeric_limits<int8_t>::digits10 + 3;
                     buf = new char[bufSize];
-                    snprintf(buf, bufSize, "%d",
-                        static_cast<int>(v[i]));
+                    snprintf(buf, bufSize, "%d", v[i]);
                 }
                 break;
             case x_uint8:
@@ -121,8 +120,7 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                     std::size_t const bufSize
                         = std::numeric_limits<uint8_t>::digits10 + 3;
                     buf = new char[bufSize];
-                    snprintf(buf, bufSize, "%u",
-                        static_cast<unsigned int>(v[i]));
+                    snprintf(buf, bufSize, "%u", v[i]);
                 }
                 break;
             case x_int16:
@@ -134,8 +132,7 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                     std::size_t const bufSize
                         = std::numeric_limits<int16_t>::digits10 + 3;
                     buf = new char[bufSize];
-                    snprintf(buf, bufSize, "%d",
-                        static_cast<int>(v[i]));
+                    snprintf(buf, bufSize, "%d", v[i]);
                 }
                 break;
             case x_uint16:
@@ -147,8 +144,7 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                     std::size_t const bufSize
                         = std::numeric_limits<uint16_t>::digits10 + 3;
                     buf = new char[bufSize];
-                    snprintf(buf, bufSize, "%u",
-                        static_cast<unsigned int>(v[i]));
+                    snprintf(buf, bufSize, "%u", v[i]);
                 }
                 break;
             case x_int32:
@@ -160,8 +156,7 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                     std::size_t const bufSize
                         = std::numeric_limits<int32_t>::digits10 + 3;
                     buf = new char[bufSize];
-                    snprintf(buf, bufSize, "%d",
-                        static_cast<int>(v[i]));
+                    snprintf(buf, bufSize, "%d", v[i]);
                 }
                 break;
             case x_uint32:
@@ -173,8 +168,7 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                     std::size_t const bufSize
                         = std::numeric_limits<uint32_t>::digits10 + 3;
                     buf = new char[bufSize];
-                    snprintf(buf, bufSize, "%u",
-                        static_cast<unsigned int>(v[i]));
+                    snprintf(buf, bufSize, "%u", v[i]);
                 }
                 break;
             case x_int64:
